chore(Build): AND-1486 Move zxing
