+++ zxing/src/main/java/piuk/blockchain/android/ui/zxing/Contents.java
