+++ zxing/src/main/java/piuk/blockchain/android/ui/zxing/FinishListener.java
