+++ zxing/src/main/java/piuk/blockchain/android/ui/zxing/InactivityTimer.java
