+++ zxing/src/main/java/piuk/blockchain/android/ui/zxing/Intents.java
