+++ zxing/src/main/java/piuk/blockchain/android/ui/zxing/camera/PreviewCallback.java
