+++ zxing/src/main/java/piuk/blockchain/android/ui/zxing/camera/exposure/FroyoExposureInterface.java
