+++ zxing/src/main/java/piuk/blockchain/android/ui/zxing/camera/open/DefaultOpenCameraInterface.java
