+++ zxing/src/main/java/piuk/blockchain/android/ui/zxing/camera/open/OpenCameraInterface.java
