+++ zxing/src/main/java/piuk/blockchain/android/ui/zxing/camera/open/OpenCameraManager.java
