+++ zxing/src/main/java/piuk/blockchain/android/ui/zxing/common/PlatformSupportManager.java
