+++ zxing/src/main/java/piuk/blockchain/android/ui/zxing/common/executor/AsyncTaskExecManager.java
