+++ zxing/src/main/java/piuk/blockchain/android/ui/zxing/encode/QRCodeEncoder.java
