+++ zxing/src/main/res/layout/activity_scan.xml
