+++ zxing/src/main/res/menu/menu_scan.xml
