+++ zxing/src/main/res/values/string.xml
@@ -1,0 +1,4 @@
+<?xml version="1.0" encoding="utf-8"?>
+<resources>
+    <string name="scan_qr">Scan QR Code</string>
+</resources>
