+++ zxing/src/main/res/values/zxing-ids.xml
